requirements.txt
flask
pandas
openpyxl
gunicorn
werkzeug
psycopg2-binary
python-dotenv
